[PONG:BASE] Refactor: Ball logic separated from circle logic.
Ball-related logic, including all IPhysicsObject implementations, extracted from Circle and moved to Ball class. Circle is now exclusively concerned with circle rendering, and Ball is an extension of Circle concerned with ball mechanics.
