+++ pong/src/pong.ts
@@ -4,7 +4,7 @@ import { showError } from "./utils.js";
 import { AspectRatio } from "./types.js";
 import { Transform } from "./transform.js";
 import { Rectangle } from "./rectangle.js";
-import { Circle } from "./circle.js";
+import { Ball } from "./ball.js";
 import { Score } from "./score.js";
 import { VerticalDashedLine } from "./net.js";
 import { onKeyDown, onKeyUp } from "./input.handlers.js";
@@ -47,7 +47,7 @@ class Pong
     private net: VerticalDashedLine;
     private leftPaddle: Rectangle;
     private rightPaddle: Rectangle;
-    private ball: Circle;
+    private ball: Ball;
     private leftScore: Score;
     private rightScore: Score;
     private drawables: IDrawable[];
@@ -65,7 +65,7 @@ class Pong
         this.net = new VerticalDashedLine(new Transform( {x: Math.round(this.canvas.width * 0.5), y: Math.round(this.canvas.height * 0.5) }, 1), "black", 5, this.canvas.height, 10);
         this.leftPaddle = new Rectangle(new Transform({ x: 100, y: Math.round(this.canvas.height * 0.5) }, 1), "black", 10, 100, 5, true);
         this.rightPaddle = new Rectangle(new Transform({ x: this.canvas.width - 100, y: Math.round(this.canvas.height * 0.5) }, 1), "black", 10, 100, 5, true);
-        this.ball = new Circle(new Transform({ x: Math.round(this.canvas.width * 0.5), y: Math.round(this.canvas.height * 0.5) }, 1), "white", 1, 10, true);
+        this.ball = new Ball(new Transform({ x: Math.round(this.canvas.width * 0.5), y: Math.round(this.canvas.height * 0.5) }, 1), "white", 1, 10, true);
         this.leftScore = new Score(new Transform({ x: Math.round(this.canvas.width * 0.25), y: Math.round(this.canvas.height * 0.2) }, 1), "0", "white", 75);
         this.rightScore = new Score(new Transform({ x: Math.round(this.canvas.width * 0.75), y: Math.round(this.canvas.height * 0.2) }, 1), "0", "white", 75);
         this.drawables = [ this.net, this.leftPaddle, this.rightPaddle, this.ball, this.leftScore, this.rightScore ];
